Панель Предупреждения PVS

+++ sources/Panel/src/Display/Colors.cpp
@@ -177,7 +177,7 @@ void ColorType::SetBrightness(float bright)
     }
     else
     {
-        int delta = (int)((bright + 0.0005F) * 100.0F) - (int)(brightness * 100.0F);
+        int delta = static_cast<int>((bright + 0.0005F) * 100.0F) - static_cast<int>(brightness * 100.0F);
 
         if (delta > 0)
         {
@@ -246,7 +246,7 @@ void ColorType::CalcSteps()
 
 void ColorType::SetColor()
 {
-    COLOR(color.value) = MAKE_COLOR((int)red, (int)green, (int)blue);
+    COLOR(color.value) = MAKE_COLOR(red, green, blue);
 }
 
 
@@ -258,7 +258,7 @@ void ColorType::ComponentChange(int delta)
 
     if (index >= 1 && index <= 3)
     {
-        *(pointers[index]) += (float)Math::Sign(delta);
+        *(pointers[index]) += static_cast<float>(Math::Sign(delta));
         Math::Limitation<float>(pointers[index], 0.0F, maxs[index]);
     }
